+++ pkg/integrations/execution/v4/config/migrate.go
@@ -2,12 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 // Contains all the bits and pieces we need to parse and manage
 // the external configuration
-package migrate
+package config
 
 import (
 	"fmt"
-	"github.com/newrelic/infrastructure-agent/pkg/integrations/execution/v3/config"
-	config2 "github.com/newrelic/infrastructure-agent/pkg/integrations/execution/v4/config"
+	v3config "github.com/newrelic/infrastructure-agent/pkg/integrations/execution/v3/config"
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 	"io/ioutil"
@@ -38,14 +37,14 @@ func V3toV4(pathConfiguration string, pathDefinition string, pathOutput string, 
 	}
 
 	// Reading old Definition file
-	v3Definition := config.Plugin{}
+	v3Definition := v3config.Plugin{}
 	err := ReadAndUnmarshallConfig(pathDefinition, &v3Definition)
 	if err != nil {
 		return fmt.Errorf("error reading old config definition: %w", err)
 	}
 
 	// Reading old Configuration file
-	v3Configuration := config.PluginInstanceWrapper{}
+	v3Configuration := v3config.PluginInstanceWrapper{}
 	err = ReadAndUnmarshallConfig(pathConfiguration, &v3Configuration)
 	if err != nil {
 		return fmt.Errorf("error reading old config configuration: %w", err)
@@ -81,7 +80,7 @@ func ReadAndUnmarshallConfig(path string, out interface{}) error {
 	return nil
 }
 
-func PopulateV4Config(v3Definition config.Plugin, v3Configuration config.PluginInstanceWrapper) (*config2.YAML, error) {
+func PopulateV4Config(v3Definition v3config.Plugin, v3Configuration v3config.PluginInstanceWrapper) (*YAML, error) {
 	if v3Configuration.IntegrationName != v3Definition.Name {
 		return nil, fmt.Errorf("IntegrationName != Name: %s!=%s", v3Configuration.IntegrationName, v3Definition.Name)
 	}
@@ -91,7 +90,7 @@ func PopulateV4Config(v3Definition config.Plugin, v3Configuration config.PluginI
 		log.Debugf("The old definitions had a os directive, %s. Usually it is not needed, use `when` field otherwhise", v3Definition.OS)
 	}
 
-	v4Config := &config2.YAML{}
+	v4Config := &YAML{}
 	for commandName, pluginV1Command := range v3Definition.Commands {
 		for _, pluginV1Instance := range v3Configuration.Instances {
 			if commandName == pluginV1Instance.Command {
@@ -104,8 +103,8 @@ func PopulateV4Config(v3Definition config.Plugin, v3Configuration config.PluginI
 	return v4Config, nil
 }
 
-func populateConfigEntry(pluginV1Instance *config.PluginV1Instance, pluginV1Command *config.PluginV1Command) config2.ConfigEntry {
-	configEntry := config2.ConfigEntry{}
+func populateConfigEntry(pluginV1Instance *v3config.PluginV1Instance, pluginV1Command *v3config.PluginV1Command) ConfigEntry {
+	configEntry := ConfigEntry{}
 	if len(pluginV1Command.Command) == 0 {
 		return configEntry
 	}
@@ -132,7 +131,7 @@ func populateConfigEntry(pluginV1Instance *config.PluginV1Instance, pluginV1Comm
 	return configEntry
 }
 
-func buildCLIArgs(pluginV1Command *config.PluginV1Command, configEntry *config2.ConfigEntry) {
+func buildCLIArgs(pluginV1Command *v3config.PluginV1Command, configEntry *ConfigEntry) {
 	for index, arg := range pluginV1Command.Command {
 		if index == 0 {
 			// the first arg in command is the binary name
@@ -149,7 +148,7 @@ func buildCLIArgs(pluginV1Command *config.PluginV1Command, configEntry *config2.
 	}
 }
 
-func writeOutput(v4Config *config2.YAML, pathDefinition string, pathConfiguration string, pathOutput string) error {
+func writeOutput(v4Config *YAML, pathDefinition string, pathConfiguration string, pathOutput string) error {
 	if v4Config == nil {
 		return fmt.Errorf("v4Config pointer is nil")
 	}
@@ -178,7 +177,7 @@ func writeOutput(v4Config *config2.YAML, pathDefinition string, pathConfiguratio
 	return nil
 }
 
-func writeV4Config(v4Config *config2.YAML, file *os.File) error {
+func writeV4Config(v4Config *YAML, file *os.File) error {
 	// see https://github.com/go-yaml/yaml/commit/7649d4548cb53a614db133b2a8ac1f31859dda8c
 	//yaml.FutureLineWrap() @TODO enable this
